grpc-python/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
fastapi==0.95.0
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-tools==1.51.3
h11==0.14.0
idna==3.4
mysql==0.0.3
mysqlclient==2.1.1
protobuf==4.22.1
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
